requirements.txt
motor
pytz
pyrogram
tgcrypto
aiofiles
python-dateutil
schedule
requests
Pyromod==1.5